requirements.txt
flask
flask-sqlalchemy
Flask-Pydantic==0.11.0
pytest